+++ h2integrate/control/control_strategies/system_level/test/test_slc_examples.py
@@ -187,7 +187,7 @@ def test_slc_yes_hydrogen(subtests, temp_copy_of_example):
             pytest.approx(
                 model.prob.get_val("finance_subgroup_hydrogen.LCOH", units="USD/kg"), rel=1e-6
             )
-            == 14.878096642042243
+            == 9.46456262909423
         )